move tracks endpoints
