+++ server/app.py
@@ -8,6 +8,7 @@ import bcrypt
 import base64
 import requests
 
+from server.utils import email_in_db, has_correct_password, username_in_db
 
 def get_db_connection():
     return psycopg2.connect(
@@ -22,44 +23,6 @@ app.secret_key = os.environ['SECRET_KEY']
 CORS(app, origins=["http://localhost:3000"],  supports_credentials=True)
 login_manager = flask_login.LoginManager()
 login_manager.init_app(app)
-
-
-def email_in_db(email):
-    conn = get_db_connection()
-    cur = conn.cursor()
-
-    cur.execute("SELECT email FROM users WHERE email = %s", (email,))
-    user = cur.fetchone()
-
-    conn.close()
-    cur.close()
-    return user
-
-def username_in_db(username):
-    conn = get_db_connection()
-    cur = conn.cursor()
-
-    cur.execute("SELECT username FROM users WHERE username = %s", (username,))
-    user = cur.fetchone()
-
-    conn.close()
-    cur.close()
-    return user
-
-
-def has_correct_password(email, password):
-    conn = get_db_connection()
-    cur = conn.cursor()
-
-    cur.execute("SELECT password FROM users WHERE email = %s", (email,))
-    correct_password = cur.fetchone()
-
-    conn.close()
-    cur.close()
-    encoded_password = password.encode('utf-8')
-    encoded_correct_password = correct_password[0].encode('utf-8')
-    return bcrypt.checkpw(encoded_password, encoded_correct_password)
-
 
 class User(flask_login.UserMixin):
     pass
@@ -154,89 +117,6 @@ def user_create():
     return 'success'
 
 
-@app.post("/tracks/create")
-@flask_login.login_required
-def track_create():
-    request_data = request.get_json()
-    conn = get_db_connection()
-    cur = conn.cursor()
-
-    cur.execute("SELECT projectid FROM projects WHERE projectid = %s",
-                (str(request_data["projectId"])))
-    projectid = cur.fetchone()
-    if projectid is None:
-        return 'project not found'
-
-    blob_data_bytes = request_data["blobData"].encode('ascii')
-    cur.execute("INSERT INTO blob_storage (blob_data) VALUES (%s)",
-                (blob_data_bytes,))
-    cur.execute(
-        "SELECT blobid FROM blob_storage WHERE blob_data = %s", (blob_data_bytes,))
-    blobId = cur.fetchone()
-
-    cur.execute("INSERT INTO tracks (instrumenttype, contributeremail, blobid, title, description) VALUES (%s, %s, %s, %s, %s)",
-                (request_data["instrumentType"], flask_login.current_user.id, blobId, request_data["title"], request_data["description"]))
-    cur.execute("SELECT trackid FROM tracks WHERE blobid = %s", blobId)
-    trackId = cur.fetchone()
-
-    cur.execute("INSERT INTO projecttracks (projectid, trackid, accepted) VALUES (%s, %s, false)",
-                (request_data["projectId"], trackId))
-
-    conn.commit()
-    conn.close()
-    cur.close()
-    return 'success'
-
-@app.get("/tracks/list")
-@cross_origin()
-def track_list():
-    conn = get_db_connection()
-    cur = conn.cursor()
-
-    cur.execute("SELECT tracks.trackid, blob_data, instrumenttype, accepted, tracks.title, tracks.description FROM projects join projecttracks on projects.projectid = projecttracks.projectid join tracks on projecttracks.trackid = tracks.trackid join blob_storage on (tracks.blobid = blob_storage.blobid) where projects.projectid = %s", (request.args.get("projectId"),))
-    tracks = cur.fetchall()
-
-    formatted_tracks = []
-    for row in tracks:
-        formatted_blob = row[1].tobytes().decode('ascii')
-        formatted_tracks.append(
-            {"trackId": row[0], "blobData": formatted_blob, "instrumentType": row[2], "accepted": row[3], "title": row[4], "description": row[5]})
-
-    conn.close()
-    cur.close()
-    return jsonify(formatted_tracks)
-
-
-@app.patch("/tracks/patch")
-@flask_login.login_required
-def track_patch():
-    request_data = request.get_json()
-    conn = get_db_connection()
-    cur = conn.cursor()
-
-    if request_data["accepted"] is None or request_data["trackIds"] is None:
-        return 'missing accepted or trackIds', 400
-    trackids = tuple(request_data["trackIds"])
-    cur.execute("SELECT tracks.trackid FROM projects natural join projecttracks join tracks on projecttracks.trackid = tracks.trackid WHERE owner = %s and tracks.trackid in %s and projectid = %s",
-                (flask_login.current_user.id, trackids, request_data["projectId"],))
-    acceptable_trackids = cur.fetchall()
-    if len(acceptable_trackids) == 0:
-        conn.close()
-        cur.close()
-        return 'No valid trackids provided', 400
-
-    acceptable_trackids = tuple(
-        trackid for subtuple in acceptable_trackids for trackid in subtuple)
-
-    cur.execute("UPDATE projecttracks SET accepted = %s WHERE trackid IN %s",
-                (request_data["accepted"], acceptable_trackids,))
-    conn.commit()
-
-    conn.close()
-    cur.close()
-    return 'success'
-
-
 @app.get("/users/loggedIn")
 @flask_login.login_required
 def get_user_logged_in():
@@ -283,7 +163,6 @@ def projecttracks_list():
     cur.close()
     conn.close()
     return jsonify(username)
-
 
 @app.patch("/users/patch")
 @flask_login.login_required
